requirements.txt
Flask
pymongo
boto3
python-dotenv
Werkzeug